Add files via upload
I wrote the code in this repository for my "Intro to Sentiment Analysis with Python" assignment
for CPT 127 during the summer of 2025. The program takes a csv file that contains 22 URLs from the Wayback
Machine on archive.org which contain archived versions of the Wikipedia entry for Elon Musk in six month intervals
starting in January of 2015 and continuing until June of 2025, and parses the body of the text for each entry 
from the html code. it then runs that through a Natural Language Processing pipeline for sentiment analysis
using the TextBlob library, calculating the polarity score (ranges from -1.0 to 1.0, which -1.0 signifying
very negative sentiment, 1.0 signifying very positive sentiment, and 0 signifying neutral sentiment) and
subjectivity score (ranges from 0.0 to 1.0 -- 0.0 being extremely objective and 1.0 being extremely subjective)
of each version, adding the data to a pandas DataFrame, and plotting the scores to visual the changes of each
over the timeframe analyzed. This code accompanies a write-up of my process and a discussion of the results that
is not posted here on GitHub.

The files are:
* 'sentiment_analysis_proj.py' - the main program
* 'sentiment_analysis_urls.csv' - the list of URL addresses called in the main program
* 'Interactive - sent_analy_proj_plot_interactive.py.ipynb' - a saved file of the interactive session of the main program where I plotted the graph (slightly older version of the code than the version of 'sentiment_analysis_proj.py uploaded here)